chore moved the DicomBufferCODEC into its own file and made the encoder and decoder sides use the same default encoding.
